test(ccxt): restructure tests
